tests: refactor to split large file
